requirements.txt
# LLM & Agent Framework
langchain>=0.1.0
langgraph>=0.0.20
openai>=1.10.0
langchain-openai>=0.0.5

# Database & Vector Store
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sqlalchemy>=2.0.0

# Embeddings & NLP
tiktoken>=0.5.0

# RAG - Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0

# RAG - Retrieval & Ranking
rank-bm25>=0.2.2
python-Levenshtein>=0.21.0
sentence-transformers>=2.2.0  # For cross-encoder reranking

# RAG - Evaluation
ragas>=0.1.0  # RAG evaluation framework
datasets>=2.0.0  # Required by RAGAS

# RAG - Web Search (optional)
tavily-python>=0.3.0
duckduckgo-search>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langdetect>=1.0.9  # Language detection for metadata
cachetools